changes made for improved stability with bitnet, full float32 may be removed later after more experiments

+++ src/model/modeling_llama.py
@@ -208,7 +208,7 @@ class LolcatsLlamaForCausalLM(LlamaForCausalLM):
             logits = torch.cat(logits, dim=-1)
         else:
             # another strange precision issue
-            hidden_states = hidden_states.to(torch.bfloat16)
+            # hidden_states = hidden_states.to(torch.bfloat16)
             logits = self.lm_head(hidden_states)
         logits = logits.float()
